add test classes and refactor quality control state
